+++ lib/Async/Await.d.ts
@@ -1,5 +1,6 @@
+/// <reference types="react" />
 import { Children } from '../interfaces/Children';
 interface Props extends Children {
 }
-export declare const Await: ({ children }: Props) => void;
+export declare const Await: ({ children }: Props) => import("react").ReactChildren | import("react").ReactChildren[];
 export {};