+++ lib/Functions/Reduce.d.ts
@@ -1,5 +1,6 @@
 import { Children } from '../interfaces/Children';
 interface Props extends Children {
+    data: any[];
 }
-export declare const Async: ({ children }: Props) => Promise<undefined>;
+export declare const Reduce: ({ data, children }: Props) => void;
 export {};