+++ lib/Loop/ForOf.d.ts
@@ -1,9 +1,8 @@
-/// <reference types="react" />
-import { Children } from '../interfaces/Children';
-interface Props extends Children {
+interface Props {
+    children: any;
     data: any[];
     key: keyof any[];
     value: any[keyof any[]];
 }
-export declare const ForOf: ({ children, data, key, value }: Props) => import("react").ReactChildren | import("react").ReactChildren[] | undefined;
+export declare const ForOf: ({ children, data, key, value }: Props) => any;
 export {};