+++ src/Async/Async.ts
@@ -7,4 +7,22 @@ import { Children } from '../interfaces/Children';
 
 interface Props extends Children {}
 
-export const Async = ({ children }: Props) => {};
+export const Async = async ({ children }: Props) => {
+  return [].concat(children as any).find((c: any) => {
+    if (c.type === Try) {
+      return null;
+    }
+
+    if (c.type === Await) {
+      return null;
+    }
+
+    if (c.type === Catch) {
+      return null;
+    }
+
+    if (c.type === Finally) {
+      return null;
+    }
+  });
+};