+++ src/Async/Await.ts
@@ -2,4 +2,4 @@ import { Children } from '../interfaces/Children';
 
 interface Props extends Children {}
 
-export const Await = ({ children }: Props) => {};
+export const Await = ({ children }: Props) => children;