+++ src/Loop/ForEach.ts
@@ -1,0 +1,8 @@
+interface Props {
+  data: any[];
+  children: any;
+}
+
+export const ForEach = ({ data, children }: Props) => {
+  return data.forEach((item, index) => children(item, index));
+};