+++ src/Loop/ForOf.ts
@@ -1,13 +1,15 @@
 import { Children } from '../interfaces/Children';
 
-interface Props extends Children {
+interface Props {
+  children: any;
   data: any[];
   key: keyof any[];
   value: any[keyof any[]];
 }
 
 export const ForOf = ({ children, data, key, value }: Props) => {
-  for ([key, value] of data) {
-    return children;
+  // for ([key, value] of data) {
+  for (value of data) {
+    return children(value);
   }
 };