+++ stories/Filter.stories.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import { Filter } from '../lib/react-render-actions.esm';
+
+export default {
+  title: 'Filter',
+  component: Filter
+};
+
+export const normal = () => <></>;