+++ stories/Map.stories.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import { Map } from '../lib/react-render-actions.esm';
+
+export default {
+  title: 'Map',
+  component: Map
+};
+
+export const normal = () => <></>;