+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # react-render-actions Changelog
 
 ## 1.0.0 (Soon)
+#### New Feature
+- conditions: `If`, `Else`, `Then`, `When`, `Unless`
+- switch: `Switch`, `Case`, `Default`
+- asynchronous: `Async`, `Await`, `Try`, `Catch`, `Finally`
+- loops: `For`, `ForOf`, `ForIn`, `ForEach`, `While`, `DoWhile`, `Break`, `Continue`
+- functions: `Map`, `Filter`, `Reduce`
+- timeouts: `SetTimeout`, `SetInterval`
+- promises: `Promise`, `Resolve`, `Reject`